# drop init time cost of apd from 0.530ms to 0.002ms