--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/cockroachdb/apd/v3
 
-go 1.17
+go 1.18
 
 require github.com/lib/pq v1.10.7