(see commit messages - please do not squash)