@mvdan 2024-08-06
cc @josharian :)