@mvdan 2024-08-20
Friendly ping @nvanbenschoten :) Let me know if I should do anything else to help this along. Note that https://github.com/cockroachdb/apd/pull/139 caused conflicts so I dropped the CI changes from this PR for now.