@jordanlewis 2024-11-22
Thanks for the contribution, @mvdan!

Since this is a performance-related change, could you re-use or write a benchmark for this and put the results in the PR? Additionally, since there's new use of sync.Once, and this library does live in the critical hot-path for other systems, would you be able to run a benchdiff on any benchmarks in this package and include the results as well?