@mvdan 2024-11-22
I did provide benchmarks in the commit messages with https://github.com/mvdan/benchinit, which as far as I can tell is the best way we have for now to measure init time cost.

This package has a large number of benchmarks, is there any in particular that would be relevant here? Running all through benchstat is possible but rather costly.