@jordanlewis 2024-11-22
Ah, thanks, I didn't notice. Regarding which are relevant, I'm afraid I'm not familiar enough with this library to tell you off-hand. Looking at the diff, I see the sync.Once added to a few spots - would it be possible to trace those callsites up and find some relevant benchmarks to run that touches the changed code? If that's "literally everything" maybe just run a reasonable looking subset.

Really just looking to make sure we're not going to introduce any obvious regressions. Thank you!