@mvdan 2024-11-23
Unfortunately the benchmark suite is rather vast when including the sub-benchmarks, so trying to run them all with `-count=6 -benchtime=1s` literally timed out after ten minutes, so I gave up on that route.

I took one of the benchmarks which is simpler and clearly affected by one of the changes, and indeed there's some overhead by the added `sync.Once`:

```
                   │     old     │                 new                 │
                   │   sec/op    │   sec/op     vs base                │
NumDigitsLookup-16   8.376µ ± 1%   9.564µ ± 2%  +14.19% (p=0.000 n=10)
```

For some context, the changes here reduce the init time cost by about 0.530ms, and this NumDigits benchmark shows a slowdown of about 0.001ms. I'm not sure what your tradeoffs are in terms of init time cost versus func/method call cost, but I'd say half a millisecond at init time is way too much; most Go packages use 5-20 microseconds at init time at most.