@jordanlewis 2024-11-25
Thanks @mvdan! It doesn't matter to Cockroach Labs if this package takes half a millisecond to initialize because half an ms of initialization time is a rounding error for CockroachDB initialization time. By contrast, a 15% increase is the time it takes to do decimal operations could be a much bigger problem because the database might do a lot of those in memory for a math-intensive query - it's not inconceivable that a change like this could cause a 5% or larger regression in the overall time it takes to perform a database query for certain workloads.

As a result, I don't think this solution is of interest to our team, unfortunately.

What if an environment variable or other such input could tell the library to not perform the expensive initialization steps, and then expose the expensive initialization as a public function that could be used for use cases that set the env variable? Just brainstorming...