@mvdan 2025-01-15
Fair enough. Half a millisecond is still unfortunate for downstream CLI users, where every millisecond wasted in startup really does matter. But I don't know what to suggest here given that `sync.Once` and others add a bit of overhead, beyond requesting that you please give some love to init time cost - in a way that the tradeoff is acceptable to you.